Refactored the codebase to remove redundant code and follow modular practices.
