+++ finetune.py
@@ -19,15 +19,15 @@ from sklearn.model_selection import train_test_split
 from skmultilearn.model_selection import iterative_train_test_split
 from torch.utils.data import DataLoader
 
+from odyssey.utils.utils import seed_everything
 from odyssey.data.dataset import FinetuneDataset, FinetuneMultiDataset
 from odyssey.data.tokenizer import ConceptTokenizer
 from odyssey.models.cehr_bert.model import BertFinetune, BertPretrain
 from odyssey.models.cehr_big_bird.model import BigBirdFinetune, BigBirdPretrain
-from odyssey.models.utils import (
+from odyssey.models.model_utils import (
     get_run_id,
     load_config,
     load_finetune_data,
-    seed_everything,
 )
