+++ odyssey/data/dataset.py
@@ -4,6 +4,7 @@ import random
 from typing import Any, Dict, List, Optional, Tuple, Union
 
 import pandas as pd
+
 import torch
 from torch.utils.data import Dataset
